server/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
google-generativeai
pypdf
numpy
pandas
scikit-learn
joblib
